+++ CHANGES.md
@@ -45,6 +45,9 @@
 - Improve error message when parsing several licenses in `(license)` (#6114,
   fixes #6103, @emillon)
 
+- odoc rules now about `ODOC_SYNTAX` and will re-run accordingly (#6010, fixes
+  #1117, @emillon)
+
 3.4.1 (26-07-2022)
 ------------------